requirements.txt
streamlit>=1.22,<2.0
requests>=2.28,<3.0
Pillow>=9.5,<11.0